Re-structured the components
