+++ src/VersatileTable/VersatileTable.scss
@@ -7,39 +7,25 @@
   box-shadow: 0 1px 6px rgba(32, 33, 36, 0.28);
   max-width: fit-content;
 
-  .vt-header-col {
-    display: inline-block;
-    padding: 5px;
-    border-bottom: 2px solid rgba(32, 33, 36, 0.28);
-    cursor: pointer;
-    position: relative;
+  // .vt-header-col {
+  //   display: inline-block;
+  //   padding: 5px;
+  //   border-bottom: 2px solid rgba(32, 33, 36, 0.28);
+  //   cursor: pointer;
+  //   position: relative;
+  // }
 
-    .sort-up-icon {
-      height: 10px;
-      padding-bottom: 5px;
-      position: absolute;
-      opacity: 0.2;
-    }
+  // .vt-body-row {
+  //   display: block;
+  //   width: fit-content;
+  //   max-width: fit-content;
+  //   border-bottom: 1px solid rgba(32, 33, 36, 0.28);
+  // }
 
-    .sort-down-icon {
-      height: 10px;
-      position: absolute;
-      padding-top: 7px;
-      opacity: 0.2;
-    }
-  }
-
-  .vt-body-row {
-    display: block;
-    width: fit-content;
-    max-width: fit-content;
-    border-bottom: 1px solid rgba(32, 33, 36, 0.28);
-  }
-
-  .vt-body-col {
-    display: inline-block;
-    padding: 5px;
-  }
+  // .vt-body-col {
+  //   display: inline-block;
+  //   padding: 5px;
+  // }
 }
 
 .vt-table-basic {
@@ -53,14 +39,6 @@
     display: inline-block;
     padding: 5px;
     cursor: pointer;
-
-    .sort-up-icon {
-      display: none;
-    }
-
-    .sort-down-icon {
-      display: none;
-    }
   }
 
   .vt-body-row {
