+++ src/VersatileTable/components/Body/Body.scss
@@ -1,0 +1,11 @@
+.vt-body-row {
+  display: block;
+  width: fit-content;
+  max-width: fit-content;
+  border-bottom: 1px solid rgba(32, 33, 36, 0.28);
+}
+
+.vt-body-col {
+  display: inline-block;
+  padding: 5px;
+}
