+++ src/VersatileTable/components/Body/NoRecordComponent/NoRecordComponent.scss
@@ -1,9 +1,9 @@
-.no-record {
+.vt-no-record {
   width: 98%;
   text-align: center;
   font-weight: bold;
 }
 
-.no-record-renderer {
+.vt-no-record-renderer {
   width: 98%;
 }
