+++ src/VersatileTable/components/Header/Header.scss
@@ -1,0 +1,7 @@
+.vt-header-col {
+  display: inline-block;
+  padding: 5px;
+  border-bottom: 2px solid rgba(32, 33, 36, 0.28);
+  cursor: pointer;
+  position: relative;
+}
