+++ src/VersatileTable/components/Header/SortIconComponent/SortIconComponent.scss
@@ -1,0 +1,13 @@
+.vt-sort-up-icon {
+  height: 10px;
+  padding-bottom: 5px;
+  position: absolute;
+  opacity: 0.2;
+}
+
+.vt-sort-down-icon {
+  height: 10px;
+  position: absolute;
+  padding-top: 7px;
+  opacity: 0.2;
+}
